코드 정리;api 엔드포인트 정리
